requirements.txt
streamlit
# googletrans==4.0.0-rc1
pillow
pandas
deep-translator
